+++ src/main/java/theflogat/technomancy/common/tiles/dynamos/TileBloodDynamo.java
@@ -19,9 +19,9 @@ public class TileBloodDynamo extends TileDynamoBase implements IFluidHandler {
 	@Override
 	public int extractFuel(int ener) {
 		if(liquid!=0){
-			float ratio = ((float) ener) / 80F;
+			float ratio = (ener) / 80F;
 			float val = 100F * ratio;
-			float fuelPerc = val / (float) Math.min(liquid, val);
+			float fuelPerc = val / Math.min(liquid, val);
 			liquid -= Math.min(liquid, val);
 			float fuel = 12F * fuelPerc;
 			return (int) fuel;
